+++ source/src/ports/generic_networkhandler.c
@@ -477,7 +477,7 @@ EipStatus NetworkHandlerProcessCyclic(void) {
     CheckAndHandleUdpGlobalBroadcastSocket();
     CheckAndHandleConsumingUdpSocket();
 
-    for(size_t socket = 0; socket <= highest_socket_handle; socket++) {
+    for(int socket = 0; socket <= highest_socket_handle; socket++) {
       if(true == CheckSocketSet(socket) ) {
         /* if it is still checked it is a TCP receive */
         if(kEipStatusError == HandleDataOnTcpSocket(socket) ) /* if error */
@@ -489,7 +489,7 @@ EipStatus NetworkHandlerProcessCyclic(void) {
     }
   }
 
-  for(size_t socket = 0; socket <= highest_socket_handle; socket++) {
+  for(int socket = 0; socket <= highest_socket_handle; socket++) {
     CheckEncapsulationInactivity(socket);
   }
